Add files via upload
Hangman Game

This Python Hangman game randomly selects a word for the player to guess. The player guesses letters, and if correct, the letters are revealed. For each incorrect guess, a part of a hangman is drawn. The player wins by guessing the word or loses by running out of guesses.